requirements.txt
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.1
gunicorn==19.9.0
Markdown==3.1.1
psycopg2==2.8.3
public==2019.4.13
pytz==2019.1
self==2019.4.13
Serializer==0.2.1
simplejson==3.16.0
sqlparse==0.3.0
whitenoise==4.1.3
